BatteryAlertSystem/WindSurf Battery Monitoring App/requirements.txt
psutil==5.9.5
plyer==2.1.0
pystray==0.19.4
Pillow==10.0.0
